backend/requirements.txt
python-dotenv
Flask==3.1.0
Flask-Cors==5.0.0
Jinja2==3.1.5
MarkupSafe==3.0.2
SQLAlchemy==2.0.36
Werkzeug==3.1.3
blinker==1.9.0
click==8.1.8
colorama==0.4.6
greenlet==3.1.1
itsdangerous==2.2.0
python-dotenv==1.0.1
typing-extensions==4.12.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Mako==1.3.8
alembic==1.14.0
ecdsa==0.19.0
pyasn1==0.6.1
python-jose==3.3.0
rsa==4.9
six==1.17.0
psycopg2==2.9.10
wheel==0.45.1
gunicorn==19.7.1